refactor: split chat-client.ts into types, dom, and connection modules
Break the 465-line monolith into focused modules:
- chat-types.ts: type definitions and message type constants
- chat-dom.ts: DOM element lookups with null guard
- chat-connection.ts: createChatClient() factory with dispatch map
- chat-client.ts: thin 8-line entry point

The 90-line if/else chain is replaced with a Record-based dispatch
map, and 7 module-level let variables are grouped into a single
ChatClientState object. All existing behavior is preserved including
security fixes (try/catch, CSS.escape, path validation, replaceChildren).
